tugas besar/dl/deployment/requirements.txt
streamlit==1.44.1
torch==2.2.2
numpy>=1.26,<2.0
pandas>=2.0,<2.3
PySastrawi==1.2.0
